backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pandas
numpy
duckdb
openpyxl
chardet
sqlparse
pydantic
pydantic-settings
python-multipart
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
